[test] Move TestSentryStacktrace test
